% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr_model.R
\name{LbsprLbms}
\alias{LbsprLbms}
\title{LbsprLbms class}
\value{
a list with estimation and simulation results. For the latter, dataframes with
the relative expected catch at length, actual catch, relative fished and un-fished populations are returned.
All vectors are standarised by dividing them by their total sum.
}
\description{
it wraps the LIME algorithm providing a common and similar interface to other lbms methods
}
\section{Super class}{
\code{\link[lbmstoolbox:Lbms]{lbmstoolbox::Lbms}} -> \code{LbsprLbms}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LbsprLbms-new}{\code{LbsprLbms$new()}}
\item \href{#method-LbsprLbms-prepare_catch_data}{\code{LbsprLbms$prepare_catch_data()}}
\item \href{#method-LbsprLbms-build_lht_context}{\code{LbsprLbms$build_lht_context()}}
\item \href{#method-LbsprLbms-run}{\code{LbsprLbms$run()}}
\item \href{#method-LbsprLbms-clone}{\code{LbsprLbms$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lbmstoolbox" data-topic="Lbms" data-id="df_to_unname_matrix"><a href='../../lbmstoolbox/html/Lbms.html#method-Lbms-df_to_unname_matrix'><code>lbmstoolbox::Lbms$df_to_unname_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lbmstoolbox" data-topic="Lbms" data-id="transpose"><a href='../../lbmstoolbox/html/Lbms.html#method-Lbms-transpose'><code>lbmstoolbox::Lbms$transpose()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprLbms-new"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprLbms-new}{}}}
\subsection{Method \code{new()}}{
Initialise class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprLbms$new(biological_params, catch_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{biological_params}}{biological parameters list}

\item{\code{catch_data}}{list with a long and wide catch dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprLbms-prepare_catch_data"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprLbms-prepare_catch_data}{}}}
\subsection{Method \code{prepare_catch_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprLbms$prepare_catch_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprLbms-build_lht_context"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprLbms-build_lht_context}{}}}
\subsection{Method \code{build_lht_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprLbms$build_lht_context()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprLbms-run"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprLbms-run}{}}}
\subsection{Method \code{run()}}{
Run LBSPR algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprLbms$run(units = "cms", bindwidth = 1, verbose = TRUE, simul = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprLbms-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprLbms-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprLbms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
