% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr_simulation.R
\name{LbsprSimulation}
\alias{LbsprSimulation}
\title{LbsprSimulation class}
\value{
A list of dataframes with the expected catch and fished/unfished population
}
\description{
it generates two dataframes containing information about the estimated vs real catch as well as
the estimated unfished vs fished population. It does not apply any depletion threshold for the unfished population.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{biology_params}}{list with the biological parameters of the analysed species}

\item{\code{estimates}}{dataframe with estimates from LBSPR model}

\item{\code{catch_data}}{list of long and wide dataframes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LbsprSimulation-new}{\code{LbsprSimulation$new()}}
\item \href{#method-LbsprSimulation-run}{\code{LbsprSimulation$run()}}
\item \href{#method-LbsprSimulation-clone}{\code{LbsprSimulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprSimulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprSimulation-new}{}}}
\subsection{Method \code{new()}}{
Initialise the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprSimulation$new(biology_params, estimates, catch_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{biology_params}}{list with the biological parameters of the analysed species}

\item{\code{estimates}}{dataframe with estimates from LBSPR model}

\item{\code{catch_data}}{list of long and wide dataframes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprSimulation-run"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprSimulation-run}{}}}
\subsection{Method \code{run()}}{
Generates the model fitting information (relative catch at length and relative fished and unfished population
at length)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprSimulation$run(verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LbsprSimulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LbsprSimulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LbsprSimulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
