% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbms_methods.R
\name{Lbms}
\alias{Lbms}
\title{Lbms class}
\description{
Base class providing common interface and parammeters for all algorithms
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{biological_params}}{biological parameters list}

\item{\code{explotation_params}}{exploitation parameters list}

\item{\code{catch_data}}{list with a long and wide catch dataframe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lbms-new}{\code{Lbms$new()}}
\item \href{#method-Lbms-prepare_catch_data}{\code{Lbms$prepare_catch_data()}}
\item \href{#method-Lbms-clone}{\code{Lbms$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lbms-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lbms-new}{}}}
\subsection{Method \code{new()}}{
Initialise the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lbms$new(
  biological_params = NULL,
  explotation_params = NULL,
  catch_data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{biological_params}}{biological parameters list}

\item{\code{explotation_params}}{exploitation parameters list}

\item{\code{icatch_data}}{list with a long and wide catch dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lbms-prepare_catch_data"></a>}}
\if{latex}{\out{\hypertarget{method-Lbms-prepare_catch_data}{}}}
\subsection{Method \code{prepare_catch_data()}}{
It prepares the length frequency data for the algorithm digestion
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lbms$prepare_catch_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a wide dataframe with lengths as rows and timestep as columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lbms-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lbms-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lbms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
