% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime.R
\name{LimeLbms}
\alias{LimeLbms}
\title{LimeLbms class}
\value{
a list with most LHT-related parameters that LIME needs to run

a list with estimation and simulation results. For the latter, dataframes with
the relative expected catch at length, actual catch, relative fished and un-fished populations are returned.
While catch is relative to lengths, population details are to ages given that LIME is an age-structured algorithm.
All vectors are standarised by dividing them by their total sum.
Finds the minumum, maximum, and starting values for SL50 and SL95
}
\description{
it wraps the LIME algorithm providing a common and similar interface to other lbms such as LBSPR
}
\section{Super class}{
\code{\link[lbmstoolbox:Lbms]{lbmstoolbox::Lbms}} -> \code{LimeLbms}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LimeLbms-new}{\code{LimeLbms$new()}}
\item \href{#method-LimeLbms-prepare_catch_data}{\code{LimeLbms$prepare_catch_data()}}
\item \href{#method-LimeLbms-build_lht_context}{\code{LimeLbms$build_lht_context()}}
\item \href{#method-LimeLbms-run}{\code{LimeLbms$run()}}
\item \href{#method-LimeLbms-clone}{\code{LimeLbms$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeLbms-new"></a>}}
\if{latex}{\out{\hypertarget{method-LimeLbms-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeLbms$new(biology_params, explotation_params, catch_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeLbms-prepare_catch_data"></a>}}
\if{latex}{\out{\hypertarget{method-LimeLbms-prepare_catch_data}{}}}
\subsection{Method \code{prepare_catch_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeLbms$prepare_catch_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeLbms-build_lht_context"></a>}}
\if{latex}{\out{\hypertarget{method-LimeLbms-build_lht_context}{}}}
\subsection{Method \code{build_lht_context()}}{
Wrapper over LIME's create_lht_list function that already calculate the starting SL50 AND SL95.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeLbms$build_lht_context(extra_lht_details = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeLbms-run"></a>}}
\if{latex}{\out{\hypertarget{method-LimeLbms-run}{}}}
\subsection{Method \code{run()}}{
Run LIME algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeLbms$run(lht_context, extra_run_context = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeLbms-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LimeLbms-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeLbms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
