% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime_model_fitting_builder.R
\name{LimeModelFittingBuilder}
\alias{LimeModelFittingBuilder}
\title{LimeModelFittingBuilder class}
\value{
a dataframe with the real standarised relative real and standarised expected catch
}
\description{
it generates two dataframes containing information about the estimated vs real catch as well as
the estimated unfished vs fished population. It does not apply any depletion threshold for the unfished population.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catch_data}}{list of long and wide dataframes}

\item{\code{mid_points}}{vector of modelled lengths by LIME}

\item{\code{estimated_catch}}{estimated catch three-dimensional matrix}

\item{\code{fished_population}}{estimated fished population matrix}

\item{\code{fished_population}}{estimated fished population matrix}

\item{\code{unfished_population}}{estimated unfished population matrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LimeModelFittingBuilder-new}{\code{LimeModelFittingBuilder$new()}}
\item \href{#method-LimeModelFittingBuilder-build_catch_df}{\code{LimeModelFittingBuilder$build_catch_df()}}
\item \href{#method-LimeModelFittingBuilder-clone}{\code{LimeModelFittingBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeModelFittingBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-LimeModelFittingBuilder-new}{}}}
\subsection{Method \code{new()}}{
Initialise the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeModelFittingBuilder$new(
  catch_data,
  mid_points,
  estimated_catch,
  fished_population,
  unfished_population
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catch_data}}{list of long and wide dataframes}

\item{\code{mid_points}}{vector of modelled lengths by LIME}

\item{\code{estimated_catch}}{estimated catch three-dimensional matrix}

\item{\code{fished_population}}{estimated fished population matrix}

\item{\code{fished_population}}{estimated fished population matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeModelFittingBuilder-build_catch_df"></a>}}
\if{latex}{\out{\hypertarget{method-LimeModelFittingBuilder-build_catch_df}{}}}
\subsection{Method \code{build_catch_df()}}{
Builds a catch dataframe with the relative and expected catch. Both data are standarised
(divided by their total sum)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeModelFittingBuilder$build_catch_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LimeModelFittingBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LimeModelFittingBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LimeModelFittingBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
